README: Restore mentions of two files, minor copyedits
The mentions of the files BUGS and INSTALL.OCTAVE were removed from README
in changeset bf41d7d03027 because they were not present in the repository
nor mentioned in any file except old changelogs.

It turns out those files are generated for Octave releases though not present
in the repository, so their mentions are restored in this changeset.