Include newer packages for package suggestions (bug #66945)
The command `pkg install -forge foo` with a non-existent package `foo` tries to
offer suggestions for what `foo` might be. This turned out to only work for
older packages from the Sourceforge era, and not newer packages that never
resided on Sourceforge. This in turn was because the code path for suggestions
was doing its own separate parsing of package metadata from an older URL
that did not include newer packages. Updating that URL led to this patch.

This patch consolidates some of the `pkg` functions so that they all use the
`___pkg___` struct instead of doing customized scraping and parsing of HTML.

All files below are in `scripts/pkg/private` unless otherwise indicated.

* list_forge_packages.m: Remove the existing code to download and construct
  the `__pkg__` struct, and spin it off as its own function. Call that function
  instead of having that code inline.

* get_validated_pkg_list.m: New file and function to download and construct
  the `__pkg__` struct from packages.octave.org.

* get_forge_pkg.m: Use the `__pkg__` struct from the new function, and rewrite
  most of the function to match. This removes the separate code path used for
  the suggestions, and allows it to see all the package names. Minor reduction
  of repeated code in the suggestions distance calculation.

* scripts/pkg/module.mk: Add new file to build system.