doc: Improve Texinfo for format().
* pr-output.cc: Use @w macro to avoid longer @code segment breaking across a
line.  Use @code insteaf of @samp for display of text that would be output
by Octave.  End table of floating point formats correctly, and start new
@table for options which apply to both integer and floating point.
Use @iftex to reduce the indentation of a very long line of code so that
it fits better in the manual.  Remove unnecessary @group/@end group usage.