pkg: Remove `-forge` from `pkg install -forge`
This changeset removes `-forge` from `pkg install` following this discussion:

https://octave.discourse.group/t/suggestions-for-clearer-name-besides-forge-to-use-with-pkg-command/6358/

Following this changeset, `pkg install foo` examines `foo` to see if it is
a local file, a URL, or a package name (in that order), and takes the
appropriate action.

This changeset does not remove all occurrences of `forge`; some still remain
in use with `pkg list -forge` and names of variables and private functions.


* pkg.m: Remove `-forge` option from `pkg install`, accept `pkg install -forge`
with a warning, redo the dispatch code before calling `install (...)`,
update documentation, minor copyedits.