pkg: Reduce server load when querying for packages
Instead of querying the server each time the user runs a `pkg` action,
this changeset queries it only once per Octave session and caches it for
subsequent calls in the same Octave session. This saves server bandwidth
and is also faster for the user (800 milliseconds querying it each time
vs 11 milliseconds caching it).

* get_validated_pkg_list.m: Make __pkg__ a persistent variable and return
  it on subsequent calls to the function.