Remove visibility attribute from function that is defined in header.
* libinterp/octave-value/ov-usr-fcn.h (octave::comment_list leading_comments):
Remove visibility attribute from function that is defined inline in the header.
It is not necessary to export that function from a library because its
definition will be available in all compilation units that include this header
anyway. Adding a dllimport attribute to a function with definition is an error
on Windows.