Use correct visibility attribute for function that is defined in header.
* liboctave/array/Array.h
(Array::Array (const Container<T>& a, const dim_vector& dv)): That constructor
is defined in the header immediately after the template class definition. Use
the correct visibility attribute for that case. Adding a dllimport attribute to
a function with following definition in the same translation unit results in a
warning on Windows.