Allow downstream to instantiate the Array<T> template class with API attributes.
* oct-conf-post-public.in.h: Do not pretend that the Array<T> template class is
like a library that would need a similar set of macros with visibility
attributes like actual libraries. Instead, define macros that are needed when
building liboctave and liboctinterp. Additionally, allow downstream projects to
define the OCTARRAY_API to the appropriate visibility attributes if they like to
instantiate the Array<T> template class with new types. If downstream projects
prefer to *not* build with visibility attributes, they do not need to change
anything.