qp: New flag to allow positive semidefinite problems (bug #64346)
The question of whether quadratic programming should allow positive
semidefinite problems or require strictly positive definite problems has
been open for some years and is the subject of several interconnected bug
reports. Octave allowed it in the past but removed it for bug #61672. This
in turn triggered bugs 64346, 64901, and 66765. This indicates that there is
no single solution satisfying everyone, so punting it to the user is probably
the best we can do.

This changeset adds a new optimization flag called "AllowSemidefinite"
to let the user deal with that choice. The default value is false for backward
compatibility, and to prevent bigger problems when calling qp from sqp.

qp.m: Add new optimization flag AllowSemidefinite, default value false.
Update docstring. Add validation of input options. Add BIST.