Convert scalar value classdef objects from or to struct for saving (bug #45833).
* libinterp/octave-value/ov-classdef.cc (octave_classdef::saveobj): Return the
data to be saved to the file. If a 'saveobj' method is defined in the classdef
file that is being saved, then that method is called upon, and the return value
of that method is used as the data to be saved to the file. If the previous
condition is true and the return type of 'saveobj' is not a struct or a classdef
object, then the Boolean flag 'custom_saveobj_ret_type' is set to true. The
method is written in order to be generically used independent of the file format
being used to save.
(octave_classdef::loadobj): Return the data to be loaded from a file into the
workspace. If a 'loadobj' method is defined in the classdef file that is being
saved, then that method is called upon loading, and the return value of that
method is used as the data to be loaded. Currently only works for scalar value
classdef objects.
* libinterp/octave-value/ov-classdef.h (octave_classdef::loadobj,
octave_classdef::saveobj): Add new member function declarations.