Fix calculation of "nearest integer" function (bug #67339)
The internal function x_nint in lo-mappers.h was being used to find the nearest
integer value to a given input x, using the expression 'std::floor (x + 0.5)'.
This expression was giving wrong results for x > 0.5 * flintmax, because the
ratio 0.5 / x was becoming lower than eps at that point, and that in turn was
causing incorrect rounding for odd numbers. In turn, it was causing problems
with mod() and rem() with numbers near (but still lower than) flintmax where
those functions were returning zero for mod (x-1, x) instead of x-1 itself.

* lo-mappers.h (x_nint): Use std::round instead of handrolled code.

* data.cc (Fmod, Frem): New BISTs for inputs near flintmax().