Change "nearest integer" function internals (bug #67339)
The internal function x_nint in mappers.h was being used to find the nearest
integer value to a given input x, using the expression 'std::floor (x + 0.5)'.
This expression was giving wrong results for x > 0.5 * flintmax, because the
ratio 0.5 / x was becoming lower than eps at that point, and that in turn was
causing incorrect rounding for odd numbers. In turn, it was causing problems
with mod() and rem() with numbers near (but still lower than) flintmax where
those functions were returning zero for mod (x-1, x) instead of x-1 itself.

This changeset replaces the logic of that internal function so that we are
calculating 'x - floor (x)' and comparing that difference to 0.5, instead
of calculating x + 0.5 which is more prone to premature rounding off for
large x.

* liboctave/numeric/mappers.h: Change function x_nint as described above,
  for float and double inputs.

* libinterp/corefcn/data.cc: Add BISTs for Fmod and Frem functions.