Replace 'round (x) == x' with 'isinteger (x)'
Discussion: https://octave.discourse.group/t/replacing-nearest-integer-routines-with-round/6690/7

This changeset changes various instances of the pattern 'round (x) == x'
to 'isinteger (x)', for clarity and to use math functions instead of
reimplementing that functionality each time.

The files xpow.cc and sparse-xpow.cc could not be fully changed over due to
platform differences between Mac and Linux (discussed in the URL above).
So only the safe changes are made, with a couple of FIXME notes added.